datapacking draft

+++ labs/memory_bound/data_packing/init.cpp
@@ -4,9 +4,9 @@
 S create_entry(int first_value, int second_value) {
   S entry;
 
-  entry.i = first_value;
-  entry.s = static_cast<short>(second_value);
-  entry.l = static_cast<long long>(first_value * second_value);
+  entry.i = static_cast<unsigned int>(first_value);
+  entry.s = static_cast<unsigned short>(second_value);
+  entry.l = static_cast<unsigned long long>(first_value * second_value);
   entry.d = static_cast<double>(first_value) / maxRandom;
   entry.b = first_value < second_value;
